// my-project/src/client/schema.ts
export const backgroundImageSchema = mediaAssetSchema.extend({});
export type BackgroundImage = z.infer<typeof backgroundImageSchema>;

export const cgImageSchema = mediaAssetSchema.extend({});
export type CGImage = z.infer<typeof cgImageSchema>;

export const soundEffectSchema = mediaAssetSchema.extend({});
export type SoundEffect = z.infer<typeof soundEffectSchema>;

export const bgmSchema = mediaAssetSchema.extend({});
export type BGM = z.infer<typeof bgmSchema>;

/* ------------------------------------------------------
    Character
------------------------------------------------------ */
export const characterSchema = z.object({
	id: z.string(),
	name: z.string(),
	images: z.record(z.string(), characterImageSchema),
});
export type Character = z.infer<typeof characterSchema>;

/* ------------------------------------------------------
    GameResources
------------------------------------------------------ */
export const gameResourcesSchema = z.object({
	characters: z.record(z.string(), characterSchema),
	backgroundImages: z.record(z.string(), backgroundImageSchema),
	soundEffects: z.record(z.string(), soundEffectSchema),
	bgms: z.record(z.string(), bgmSchema),
	cgImages: z.record(z.string(), cgImageSchema),
});
export type GameResources = z.infer<typeof gameResourcesSchema>;

/* ------------------------------------------------------
    GameEvent (Discriminated Union)
------------------------------------------------------ */

export const textRenderEventSchema = z.object({
	id: z.string(),
	type: z.literal("text"),
	category: z.literal("message"),
	order: z.string(),
	text: z.string().min(1, { message: "テキストを入力してください" }),
	characterName: z.string().optional(),
});

export type TextRenderEvent = z.infer<typeof textRenderEventSchema>;
// ...
export const effectType = ["fadeIn", "fadeOut", "shake"] as const;
export const effectEventSchema = z.object({
	id: z.string(),
	type: z.literal("effect"),
	category: z.literal("effect"),
	order: z.string(),
	effectType: z.enum(effectType),
	transitionDuration: z.union([z.number(), z.string().transform(Number)]),
});

export type EffectEvent = z.infer<typeof effectEventSchema>;

export const characterEffectType = [
	"shake",
	"flash",
	"bounce",
	"sway",
	"wobble",
	"blackOn",
	"blackOff",
] as const;
export const characterEffectEventSchema = z.object({
	id: z.string(),
	type: z.literal("characterEffect"),
	category: z.literal("character"),
	order: z.string(),
	characterId: z.string(),
	effectType: z.enum(characterEffectType),
	transitionDuration: z.union([z.number(), z.string().transform(Number)]),
});

export type CharacterEffectEvent = z.infer<typeof characterEffectEventSchema>;
// ...
export const choiceSchema = z.object({
	id: z.string(),
	text: z.string(),
	nextSceneId: z.string(),
});
export type Choice = z.infer<typeof choiceSchema>;
// ...
export type Stage = {
	cg: {
		item: {
			id: string;
			scale: number;
			position: [number, number];
		} | null;
		transitionDuration: number;
	};
	background: {
		id: string;
		transitionDuration: number;
	} | null;
	characters: {
		transitionDuration: number;
		items: {
			id: string;
			scale: number;
			imageId: string;
			position: [number, number];
			effect: {
				type: CharacterEffectEvent["effectType"];
				transitionDuration: number;
			} | null;
		}[];
	};
	dialog: {
		isVisible: boolean;
		text: string;
		characterName: string | undefined;
		transitionDuration: number;
	};
	choices: Choice[];
	soundEffect: {
		id: string;
		volume: number;
		loop: boolean;
		isPlaying: boolean;
		transitionDuration: number;
	} | null;
	bgm: {
		id: string;
		volume: number;
		loop: boolean;
		isPlaying: boolean;
		transitionDuration: number;
	} | null;
	effect: {
		type: EffectEvent["effectType"];
		transitionDuration: number;
	} | null;
};

export type MessageHistory = {
	text: string;
	characterName?: string;
	isChoice?: boolean;
};

export type GameState = "loading" | "beforeStart" | "playing" | "idle" | "end";

export type ResourceCache = {
	characters: {
		[id: string]: Character & {
			images: { [id: string]: { cache: HTMLImageElement } };
		};
	};
	backgroundImages: {
		[id: string]: BackgroundImage & { cache: HTMLImageElement };
	};
	cgImages: {
		[id: string]: CGImage & { cache: HTMLImageElement };
	};
	soundEffects: { [id: string]: SoundEffect & { cache: Howl } };
	bgms: { [id: string]: BGM & { cache: Howl } };
};